+++ osm-jigsaw-parser/src/main/scala/input/AreaReading.scala
@@ -1,4 +1,4 @@
-package steps
+package input
 
 import java.io.InputStream
